requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0,<0.129.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.0.0

# ── MLX — primary runtime on Apple Silicon (M1/M2/M3/M4) ────────────────────
# Installs mlx core + mlx-lm (generate, load helpers).
# Already bundled with the right Metal kernels — no CUDA/ROCm needed.
mlx>=0.22.0
mlx-lm>=0.22.0

# ── llama-cpp-python — GGUF fallback runtime ─────────────────────────────────
# Pre-built Metal wheel (no cmake needed on macOS):
#   pip install llama-cpp-python --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/metal
llama-cpp-python>=0.3.4

# ── HuggingFace — prefetch, tokenizers, embeddings ───────────────────────────
huggingface_hub>=0.22.0
transformers>=4.40.0
tokenizers>=0.19.0
sentence-transformers>=2.7.0
accelerate>=0.29.0

# ── Vector DB ─────────────────────────────────────────────────────────────────
chromadb>=1.0.0

# ── Web search & scraping ─────────────────────────────────────────────────────
ddgs>=9.0.0
trafilatura>=1.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
lxml_html_clean>=0.1.0

# ── YouTube transcript extraction ─────────────────────────────────────────────
youtube-transcript-api>=0.6.2

# ── File parsing (PDF, DOCX, images / OCR) ────────────────────────────────────
pypdf>=4.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10

# ── Knowledge graph ───────────────────────────────────────────────────────────
networkx>=3.3

# ── Config + utils ────────────────────────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0
